Refactor day 11 - Add post-transform function
In part 2 of this problem, the division by 3 after the worry value changes
is removed, which will result in the values getting too large; however,
the problem hints that there's a different way we can keep the values
small. Having a generalized post-transform function that we can apply
instead of just "divide by 3" makes it possible to insert some new method
of keeping the values small.