Solve day 11, pt 2
The solution for keeping the values of worry small is to mod them by the
product of all the prime numbers that the monkeys evaluate against (and
the next prime, 23) - this ensures that the values never get too large,
and continue to be congruent to the set of numbers that would be produced
with an infinite-precision integer system. Because math.